Fixed issue with Cat class
If you look at your original code, you'll see that you didn't put a method signature for the code; you just put curly braces around it. This creates what is called a "static initializer," which can be useful in ways I can explain in person later, but which doesn't allow the use of "this" because it's static and therefore class-scope as opposed to instance-scope. Remind me later and I can explain further if you want.